Made an automation script for download, unrar and process files. Various bug fixes throughout as well

+++ dem file utilities/hltvunrar/hltv_unrar.py
@@ -7,18 +7,24 @@ os.environ['UNRAR_LIB_PATH'] = 'C:/Program Files (x86)/UnrarDLL/UnRAR.dll'
 from unrar import rarfile
 
 
-def hltv_unrar(dl_location, processed_location):
+def hltv_unrar(dl_location='D:/CSGOProGames/!rarfiles/', processed_location='D:/CSGOProGames/!demofiles/'):
 
     if not os.path.exists(processed_location):
         os.mkdir(processed_location)
 
     for file in glob.glob(dl_location + '*.rar'):
-        rar = rarfile.RarFile(file)
+        try:
+            rar = rarfile.RarFile(file)
+        except:
+            continue
         rar.extractall(path=dl_location)
         extcode = file.split('\\')[-1].split('_')[0]
         dems = rar.namelist() #list of file names in archive (not including path)
         for dem in dems:
-            os.rename(dl_location + dem, processed_location + extcode + '_' + dem)
+            if not os.path.exists(processed_location + extcode + '_' + dem):
+                os.rename(dl_location + dem, processed_location + extcode + '_' + dem)
+        # Delete rar when I'm done with it so I can hold more uncompressed dem files
+        os.remove(file)
 
 if __name__ == '__main__':
-    hltv_unrar('D:/CSGOProGames/!rarfiles/', 'D:/CSGOProGames/!demofiles/')
+    hltv_unrar()
